requirements.txt
adapter-transformers==2.1.2
black==21.9b0
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
cymem==2.0.5
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0-py3-none-any.whl
filelock==3.2.0
fsspec==2021.9.0
huggingface-hub==0.0.17
idna==3.2
isort==5.9.3
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
murmurhash==1.0.5
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathspec==0.9.0
pathy==0.6.0
Pillow==8.3.2
platformdirs==2.3.0
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
tokenizers==0.10.3
tomli==1.2.1
torch==1.9.1
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wordcloud==1.8.1
zstandard==0.15.2
zstd==1.5.0.2
